requirements/base.txt
# django core 
django==3.0.2
django-environ==0.4.5
mysqlclient==1.4.6
Pillow==7.0.0
django-mptt==0.10.0
django-extensions==2.2.5
django-redis==4.11.0
redis==3.4.1

# celery
# kombu==4.6.3
# celery==4.3.0
# flower==0.9.3

# extensions support
qiniu==7.2.6
# weixin-python==0.5.5
# aliyun-python-sdk-core==2.13.15


# admin
django-admin-json-editor==0.2.0
django-ckeditor==5.6.1

#docs
drf-yasg==1.17.0
pycryptodome==3.9.4

# rest framework
django-filter==2.2.0
djangorestframework==3.11.0
django-crispy-forms==1.8.1

# deploy
daphne==2.3.0

# excel
# openpyxl==3.0.3
xlrd==1.2.0
